Added onUpdateMealPlan(Callbacks) but still breaks, needs to modify the 00 and the size of the TextField

+++ app/src/main/java/com/organization/elephant/data/models/MealPlan.kt
@@ -3,10 +3,10 @@ package com.organization.elephant.data.models
 data class MealPlan(
     val startDateInMillis: Long,
     val endDateInMillis: Long,
-    val carbs: Int,
-    val fat: Int,
-    val protein: Int,
-    val calories: Int,
+    val carbs: Double,
+    val fat: Double,
+    val protein: Double,
+    val calories: Double,
     val mealGroups:List<MealGroup>
 ){
     var formattedStartDate:String = "01/02/1200"
